Use secure wss connection, set server to only listen on localhost
Thus to connect to the ws server, clients need to either be on localhost
or go through the reverse proxy in nginx which now requires SSL. Thus
it is a wss server and now clients needn't know the port number at all,
it is 443 from their point of view. This also means that I can close the
old port in the firewall.